+++ Tests/SwiftStdlibTests/SequenceExtensionsTests.swift
@@ -134,6 +134,16 @@ final class SequenceExtensionsTests: XCTestCase {
         XCTAssertEqual(["James", "Wade", "Bryant"].sum(for: \.count), 15)
         XCTAssertEqual(["a", "b", "c", "d"].sum(for: \.count), 4)
     }
+    
+    func testProduct() {
+        XCTAssertEqual([1, 2, 3, 4, 5].product(), 120)
+        XCTAssertEqual([1.2, 2.3, 3.4, 4.5, 5.6].product(), 236.4768, accuracy: 0.001)
+    }
+
+    func testMapFunctionProduct() {
+        XCTAssertEqual(["James", "Wade", "Bryant"].product(for: \.count), 120)
+        XCTAssertEqual(["a", "b", "c", "d"].product(for: \.count), 1)
+    }
 
     func testKeyPathSorted() {
         let array = ["James", "Wade", "Bryant"]